typing: Define events in a separate module, import explicitly
