+++ beets/plugins.py
@@ -10,15 +10,7 @@ from collections import defaultdict
 from functools import cached_property, wraps
 from importlib import import_module
 from pathlib import Path
-from typing import (
-    TYPE_CHECKING,
-    Any,
-    ClassVar,
-    Literal,
-    TypedDict,
-    TypeVar,
-    overload,
-)
+from typing import TYPE_CHECKING, Any, ClassVar, TypeVar, overload
 
 import mediafile
 from typing_extensions import Never, ParamSpec, Unpack
@@ -29,18 +21,60 @@ from beets.util import unique_list
 from beets.util.deprecation import deprecate_for_maintainers, deprecate_for_user
 
 if TYPE_CHECKING:
-    from collections.abc import Callable, Iterable, Iterator, Mapping, Sequence
+    from collections.abc import Callable, Iterable, Iterator, Sequence
 
     from confuse import Subview
 
     from beets.autotag import AlbumInfo, TrackInfo
-    from beets.autotag.match import AlbumMatch
     from beets.dbcore.db import FieldQueryType
     from beets.dbcore.types import Type
     from beets.importer import Action, ImportSession, ImportTask
-    from beets.library import Album, Item, LibModel, Library
+    from beets.library import Album, Item, Library
     from beets.ui import Subcommand
     from beets.util import PromptChoice
+
+    from .events import (
+        AfterConvertEventArgs,
+        AfterConvertEventType,
+        AfterWriteEventArgs,
+        AfterWriteEventType,
+        AlbumEventArgs,
+        AlbumEventType,
+        AlbumImportedEventArgs,
+        AlbumImportedEventType,
+        AlbumInfoReceivedEventArgs,
+        AlbumInfoReceivedEventType,
+        AlbumMatchedEventArgs,
+        AlbumMatchedEventType,
+        BeforeChooseCandidateEventType,
+        DatabaseChangeEventArgs,
+        DatabaseChangeEventType,
+        EventType,
+        ImportBeginEventArgs,
+        ImportBeginEventType,
+        ImportEventArgs,
+        ImportEventType,
+        ImportTaskBeforeChoiceEventType,
+        ImportTaskCreatedEventType,
+        ImportTaskEventArgs,
+        ImportTaskEventType,
+        ItemEventArgs,
+        ItemEventType,
+        ItemImportedEventArgs,
+        ItemImportedEventType,
+        ItemPathEventArgs,
+        ItemPathEventType,
+        LibraryEventArgs,
+        LibraryEventType,
+        MetadataReceivedEventType,
+        MusicBrainzExtractEventArgs,
+        MusicBrainzExtractEventType,
+        NoArgsEventType,
+        TrackInfoReceivedEventArgs,
+        TrackInfoReceivedEventType,
+        WriteEventArgs,
+        WriteEventType,
+    )
 
     # TYPE_CHECKING guard is needed for any derived type
     # which uses an import from `beets.library` and `beets.imported`
@@ -60,146 +94,6 @@ PLUGIN_NAMESPACE = "beetsplug"
 
 # Plugins using the Last.fm API can share the same API key.
 LASTFM_KEY = "2dc3914abf35f0d9c92d97d8f8e42b43"
-
-AfterWriteEventType = Literal["after_write"]
-ItemPathEventType = Literal[
-    "before_item_moved",
-    "item_copied",
-    "item_hardlinked",
-    "item_linked",
-    "item_moved",
-    "item_reflinked",
-]
-BeforeChooseCandidateEventType = Literal["before_choose_candidate"]
-ImportTaskBeforeChoiceEventType = Literal["import_task_before_choice"]
-ImportTaskCreatedEventType = Literal["import_task_created"]
-ImportTaskEventType = Literal[
-    "import_task_apply",
-    "import_task_choice",
-    "import_task_files",
-    "import_task_start",
-]
-ImportEventType = Literal["import"]
-AlbumImportedEventType = Literal["album_imported"]
-AlbumEventType = Literal["album_removed", "art_set"]
-AlbumInfoReceivedEventType = Literal["albuminfo_received"]
-TrackInfoReceivedEventType = Literal["trackinfo_received"]
-MetadataReceivedEventType = (
-    AlbumInfoReceivedEventType | TrackInfoReceivedEventType
-)
-AlbumMatchedEventType = Literal["album_matched"]
-LibraryEventType = Literal["cli_exit", "library_opened"]
-DatabaseChangeEventType = Literal["database_change"]
-ImportBeginEventType = Literal["import_begin"]
-ItemImportedEventType = Literal["item_imported"]
-ItemEventType = Literal["item_removed"]
-WriteEventType = Literal["write"]
-MusicBrainzExtractEventType = Literal["mb_album_extract", "mb_track_extract"]
-NoArgsEventType = Literal["pluginload", "smartplaylist_update"]
-AfterConvertEventType = Literal["after_convert"]
-EventType = (
-    AfterWriteEventType
-    | ItemPathEventType
-    | BeforeChooseCandidateEventType
-    | ImportTaskBeforeChoiceEventType
-    | ImportTaskCreatedEventType
-    | ImportTaskEventType
-    | ImportEventType
-    | AlbumImportedEventType
-    | AlbumEventType
-    | MetadataReceivedEventType
-    | AlbumMatchedEventType
-    | LibraryEventType
-    | DatabaseChangeEventType
-    | ImportBeginEventType
-    | ItemImportedEventType
-    | ItemEventType
-    | WriteEventType
-    | MusicBrainzExtractEventType
-    | NoArgsEventType
-    | AfterConvertEventType
-)
-
-
-class AfterWriteEventArgs(TypedDict):
-    item: Item
-    path: bytes
-
-
-class ItemPathEventArgs(TypedDict):
-    item: Item
-    source: bytes
-    destination: bytes
-
-
-class ImportTaskEventArgs(TypedDict):
-    session: ImportSession
-    task: ImportTask
-
-
-class ImportEventArgs(TypedDict):
-    lib: Library
-    paths: list[bytes]
-
-
-class AlbumImportedEventArgs(TypedDict):
-    lib: Library
-    album: Album
-
-
-class AlbumEventArgs(TypedDict):
-    album: Album
-
-
-class AlbumInfoReceivedEventArgs(TypedDict):
-    info: AlbumInfo
-
-
-class TrackInfoReceivedEventArgs(TypedDict):
-    info: TrackInfo
-
-
-class AlbumMatchedEventArgs(TypedDict):
-    match: AlbumMatch
-
-
-class LibraryEventArgs(TypedDict):
-    lib: Library
-
-
-class DatabaseChangeEventArgs(TypedDict):
-    lib: Library
-    model: LibModel
-
-
-class ImportBeginEventArgs(TypedDict):
-    session: ImportSession
-
-
-class ItemImportedEventArgs(TypedDict):
-    lib: Library
-    item: Item
-
-
-class ItemEventArgs(TypedDict):
-    item: Item
-
-
-class WriteEventArgs(TypedDict):
-    item: Item
-    path: bytes
-    tags: dict[str, Any]
-
-
-class MusicBrainzExtractEventArgs(TypedDict):
-    data: Mapping[str, Any]
-
-
-class AfterConvertEventArgs(TypedDict):
-    item: Item
-    dest: bytes
-    keepnew: bool
-
 
 # Global logger.
 log = logging.getLogger("beets")
